Update patch set 6
Patch Set 6:

I think the problem is because mariadb supports mysql as a param so i carn't fix that as that would be down to the plugin. But as far as my testing is, it works. I can tell the difference as the session variables work in the mariadb connector copared to the mysql one which dosent (though the session variables are supported in the mysql connector it dosent support change the character encoding and collation like utf8 to utf8mb4 but this works in mariadb connector)

Patch-set: 6
Reviewer: Gerrit User 1029953 <1029953@173816e5-2b9a-37c3-8a2e-48639d4f1153>