Update patch set 3
Patch Set 3:

> Patch Set 2: Code-Review-1
> 
> IIUC this wasn't done originally because StandardKeyEncoder doesn't work with GWT, so this change would effectively remove GWT support from gwtorm. 
> 
> Maybe this is ok because gwtorm is EOL, but if so, I would like to be more intentional about this change: document it, and possibly even start deleting GWT-only code from this project.
> 
> (If you can prove I'm wrong, e.g. by running an old version of Gerrit with the GWT UI against a version of gwtorm with this change patched in, then I rescind my objection and wish we had done this years ago.)

I  checked the StandardKeyEncoder implementation and figured, that only dependencies from JDK are used, so my assumption was: it should work with GWT application. But, trying to use it in gerrit, GWT compilation was failing for that class: StandardKeyEncoder

  No source code is available for type com.google.gwtorm.server.StandardKeyEncoder; did you forget to inherit a required module?

This is because the GWT module that gwtorm is exposing only contains client package, while StandardKeyEncoder is located on server package.

The fix is trivial, just move StandardKeyEncoder to client package.

Done.

Patch-set: 3
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>